requirements.txt
# Text Cloud
streamlit
pandas
matplotlib
wordcloud
nltk
spacy
jieba
googletrans
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0.tar.gz

# LangChain
python-dotenv
markitdown
langchain-ollama
langchain-google-genai
chromadb
langchain-chroma
langchain-core